refactor: 오목 데이터를 초기화 시키는 기능 별도 클래스로 분리
